+++ EssentialFeed/EssentialFeed/Image Comments API/ImageCommentsMapper.swift
@@ -7,7 +7,7 @@
 
 import Foundation
 
-final class ImageCommentsMapper {
+public final class ImageCommentsMapper {
     private struct Root:Decodable {
         private let items : [Item]
         
@@ -29,7 +29,7 @@ final class ImageCommentsMapper {
         
     private init() { }
     
-    internal static func map(_ data:Data, from response:HTTPURLResponse) throws -> [ImageComment] {
+    public static func map(_ data:Data, from response:HTTPURLResponse) throws -> [ImageComment] {
         let decoder = JSONDecoder()
         decoder.dateDecodingStrategy = .iso8601
