+++ EssentialFeed/EssentialFeedTests/Image Comments API/ImageCommentsMapperTests.swift
@@ -1,0 +1,85 @@
+//
+//  ImageCommentsMapperTests.swift
+//  EssentialFeedTests
+//
+//  Created by Hitesh on 28/02/22.
+//
+
+import XCTest
+import EssentialFeed
+
+class ImageCommentsMapperTests: XCTestCase {
+    
+    func test_map_throwsErrorOnNon2xxHTTPResponse() throws {
+        let json = makeItemJSON([])
+        let sample = [199, 150, 300, 400, 500]
+        
+        try sample.forEach { code in
+            XCTAssertThrowsError(
+                try ImageCommentsMapper.map(json, from: HTTPURLResponse(statusCode: code))
+            )
+        }
+    }
+    
+    func test_map_throwsErrorOn2xxHTTPResponseWithInvalidJSON() throws {
+        let invalidJSON = Data("invalid json".utf8)
+        let sample = [200, 201, 250, 280, 299]
+
+        try sample.forEach { code in
+            XCTAssertThrowsError(
+                try ImageCommentsMapper.map(invalidJSON, from: HTTPURLResponse(statusCode: code))
+            )
+        }
+    }
+    
+    func test_map_deliversNoItemsOn2xxHTTPResponseWithEmptyJSONList() throws {
+        let emptyJOSN = makeItemJSON([])
+        let sample = [200, 201, 250, 280, 299]
+
+        try sample.forEach { code in
+            let result = try ImageCommentsMapper.map(emptyJOSN, from: HTTPURLResponse(statusCode: code))
+            
+            XCTAssertEqual(result, [])
+        }
+    }
+    
+    func test_load_deliversItemsOn2xxHTTPResponseWithInvalidJSON() throws {
+        let item1 = makeItem(
+            id: UUID(),
+            message: "a message",
+            createdAt: (date: Date(timeIntervalSince1970: 1598627222), iso8601String: "2020-08-28T15:07:02+00:00"),
+            username: "a username"
+            )
+        
+        let item2 = makeItem(
+            id: UUID(),
+            message: "another message",
+            createdAt: (date: Date(timeIntervalSince1970: 1577881882), iso8601String: "2020-01-01T12:31:22+00:00"),
+            username: "another username"
+            )
+        
+        let json = makeItemJSON([item1.json, item2.json])
+        let sample = [200, 201, 250, 280, 299]
+
+        try sample.forEach { code in
+            let result = try ImageCommentsMapper.map(json, from: HTTPURLResponse(statusCode: code))
+            
+            XCTAssertEqual(result, [item1.model,item2.model])
+        }
+    }
+    
+    // MARK: - Helpers
+
+    private func makeItem(id: UUID, message: String, createdAt: (date: Date, iso8601String: String), username: String) -> (model: ImageComment,json:[String:Any]) {
+        let model = ImageComment(id: id, message: message, createdAt: createdAt.date, username: username)        
+        let itemJSON: [String:Any] = [
+            "id": id.uuidString,
+            "message": message,
+            "created_at": createdAt.iso8601String,
+            "author": [
+                "username": username
+            ]
+        ]
+        return (model, itemJSON)
+    }
+}
